requirements.txt
streamlit==1.33.0
rembg==2.0.50
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
